refactor + create selectors module
