+++ src/redux-modules/uiSlice.ts
@@ -3,7 +3,6 @@ import { createSlice } from "@reduxjs/toolkit";
 import { RootState } from "./store";
 import { fetchPowerControlInfo } from "./thunks";
 import { PowerControlInfo } from "../utils/constants";
-import { Devices } from "../backend/utils";
 
 type UiStateType = {
   powerControlInfo?: PowerControlInfo;
@@ -34,22 +33,6 @@ export const uiSlice = createSlice({
 
 export const selectPowerControlInfo = (state: RootState) => {
   return state.ui.powerControlInfo;
-};
-
-export const selectDeviceName = (state: RootState) => {
-  return state.ui.powerControlInfo?.deviceName || "";
-};
-
-export const selectIsSteamDeck = (state: RootState) => {
-  const deviceName = selectDeviceName(state);
-
-  if (
-    deviceName.includes(Devices.STEAM_DECK_LCD) ||
-    deviceName.includes(Devices.STEAM_DECK_OLED)
-  ) {
-    return true;
-  }
-  return false;
 };
 
 export const selectScalingDriver = (state: RootState) => {
