+++ src/utils/selectors.tsx
@@ -1,0 +1,18 @@
+import { RootState } from "../redux-modules/store";
+import { Devices } from "../backend/utils";
+
+export const selectDeviceName = (state: RootState) => {
+  return state.ui.powerControlInfo?.deviceName || "";
+};
+
+export const selectIsSteamDeck = (state: RootState) => {
+  const deviceName = selectDeviceName(state);
+
+  if (
+    deviceName.includes(Devices.STEAM_DECK_LCD) ||
+    deviceName.includes(Devices.STEAM_DECK_OLED)
+  ) {
+    return true;
+  }
+  return false;
+};
